Gross or Net metadata information

A string list containing values indicating whether the output of the functions Recca::finaldemand_aggregates, PFUWorkflow::calculate_fu_ex_total, PFUWorkflow::calculate_fu_ex_product, PFUWorkflow::calculate_fu_ex_sector, and PFUWorkflow::calculate_finaluseful_ex_data are in Gross or Net terms. To be supplied to the metadata columns PFUWorkflow::sea_cols$gross_net_colname.

Usage

gross_net_metadata

Format

A string list with 2 entries.

gross_value

The string "Gross" indicating that final demand was calculated for both EIOU and non-EIOU sectors. See Recca::finaldemand_aggregates.

net_value

The string "Net" indicating that final demand was calculated for only non-EIOU sectors. See Recca::finaldemand_aggregates.

Examples

gross_net_metadata
#> $gross_value
#> [1] "Gross"
#> 
#> $net_value
#> [1] "Net"
#>