Load a country concordance table

In addition to loading the country concordance table, this function checks for errors: missing 3-letter country codes and repeated 3-letter country codes. Both of these issues will cause a problem.

load_country_concordance_table(
  country_concordance_path,
  sheet = "country_concordance_table",
  pfu_code_colname = "PFU.code"
)

Arguments

country_concordance_path

The path to the country concordance file. The file is assumed to be an Excel file.

sheet

The name of the sheet to be read. Default is "country_concordance_table".

pfu_code_colname

The name of the column in the country concordance table that contains 3-letter country codes to be used in the workflow. Default is "PFU.code".

Value

A country concordance table.