A map for aggregating products

A string list containing a map for aggregating energy products.

product_aggregation_map

Format

A string list with 3 entries.

Oil and oil products

All energy products counted as oil and oil products.

Natural gas

All energy products counted as natural gas.

Renewables

All energy products counted as renewables.

Biofuels and waste products

All energy products counted as biofuels and waste products.

Examples

product_aggregation_map
#> $`Oil and oil products`
#>  [1] "Crude/NGL/feedstocks (if no detail)"  
#>  [2] "Crude oil"                            
#>  [3] "Natural gas liquids"                  
#>  [4] "Additives/blending components"        
#>  [5] "Other hydrocarbons"                   
#>  [6] "Oil shale and oil sands"              
#>  [7] "Refinery feedstocks"                  
#>  [8] "Refinery gas"                         
#>  [9] "Ethane"                               
#> [10] "Liquefied petroleum gases (LPG)"      
#> [11] "Motor gasoline excl. biofuels"        
#> [12] "Aviation gasoline"                    
#> [13] "Gasoline type jet fuel"               
#> [14] "Kerosene type jet fuel excl. biofuels"
#> [15] "Other kerosene"                       
#> [16] "Gas/diesel oil excl. biofuels"        
#> [17] "Fuel oil"                             
#> [18] "Naphtha"                              
#> [19] "White spirit & SBP"                   
#> [20] "Lubricants"                           
#> [21] "Bitumen"                              
#> [22] "Paraffin waxes"                       
#> [23] "Petroleum coke"                       
#> [24] "Other oil products"                   
#> 
#> $`Natural gas`
#> [1] "Natural gas"
#> 
#> $Renewables
#> [1] "Geothermal"           "Hydro"                "Solar photovoltaics" 
#> [4] "Solar thermal"        "Tide, wave and ocean" "Wind"                
#> [7] "Other sources"       
#> 
#> $`Biofuels and waste products`
#>  [1] "Industrial waste"                        
#>  [2] "Municipal waste (renewable)"             
#>  [3] "Municipal waste (non-renewable)"         
#>  [4] "Primary solid biofuels"                  
#>  [5] "Biogases"                                
#>  [6] "Biogasoline"                             
#>  [7] "Biodiesels"                              
#>  [8] "Bio jet kerosene"                        
#>  [9] "Other liquid biofuels"                   
#> [10] "Non-specified primary biofuels and waste"
#> [11] "Charcoal"                                
#>