Save the PSUT target in a pinboard.

Releases (release = TRUE) or not (relase = FALSE) a new version of the PSUT target using the pins package.

release_psut(
  pipeline_releases_folder,
  targ,
  targ_name,
  type = "rds",
  release = FALSE
)

Arguments

pipeline_releases_folder

The folder that contains the pinboard for releases from the pipeline.

targ

The target object to be saved to the pinboard.

targ_name

The name of the target object.

type

The type of the target. Default is "rds".

release

A boolean telling whether to do a release. Default is FALSE.

Value

If release is TRUE, the fully-qualified path name of the targ file in the pinboard. If release is FALSE, NULL.

Details

Released versions of the PSUT target can be obtained as shown in examples.

Examples

if (FALSE) {
# Establish the pinboard
pinboard <- pins::board_folder("~/Dropbox/Fellowship 1960-2015 PFU database/PipelineReleases/")
# Get information about the `PSUT` target in the pinboard
pinboard %>%
  pins::pin_meta(name = "psut")
# Find versions of the `PSUT` target
pinboard %>%
  pins::pin_versions(name = "psut")
# Get the latest copy of the `PSUT` target.
my_psut <- pinboard %>%
  pins::pin_read(name = "psut")
# Retrieve a previous version of the `PSUT` target.
my_old_psut <- pinboard %>%
  pins::pin_read(name = "psut", version = "20220218T023112Z-1d9e1")}