PFUWorkflow data frame column names

A string list containing named names of columns in PFUWorkflow data frames. The data frames can be tidy (with one row for each data point) or wide (with years spread to the right). Items in the list act to compliment the column names in IEATools::iea_cols.

Usage

sea_cols

Format

A string list with 13 entries.

stage_colname

The name of a metadata column containing the stage of the energy conversion chain, usually "Primary", "Final", or "Useful".

gross_net_colname

The name of a metadata column containing information as to whether aggregated data at the final and useful stage is in "Gross" or "Net" terms, see Recca::finaldemand_aggregates() and Recca::primary_aggregates().

e_product

The name of a metadata column containing the names of energy products.

sector_colname

The name of a metadata column containing the names of final demand sectors.

flow_colname

The name of a metadata column containing the names of primary flows.

agg_by_colname

The name of a column containing the variable by which data was aggregated. Usually using Recca::finaldemand_aggregates() and Recca::primary_aggregates(), and usually one of "Flow", "Sector", "Product", or "Total".

fd_sectors_colname

The name of a column containing the list of final demand sectors desired for analysis. Usually created by PFUWorkflow::get_fd_sectors() and PFUWorkflow::create_fd_sectors_list().

p_ind_comp_colname

The name of a column containing lists of primary industries desired for analysis. Usually created by using Recca::find_p_industry_names().

p_ind_prefix_colname

The name of a column containing the list of primary industry prefixes desired for analysis. Usually supplied to Recca::find_p_industry_names() to return p_ind_comp.

ex_colname

The name of a column containing energy or exergy data.

ex_p_colname

The name of a column containing energy or exergy data at the primary stage. Usually produced by Recca::primary_aggregates().

ex_net_colname

The name of a column containing energy or exergy data at the final and/or useful stage and in net terms. Usually produced by Recca::finaldemand_aggregates().

ex_gross_colname

The name of a column containing energy or exergy data at the final and/or useful stage and in gross terms. Usually produced by Recca::finaldemand_aggregates().

Examples

sea_cols
#> $stage_colname
#> [1] "Stage"
#> 
#> $gross_net_colname
#> [1] "Gross.Net"
#> 
#> $e_product_colname
#> [1] "E.product"
#> 
#> $sector_colname
#> [1] "Sector"
#> 
#> $flow_colname
#> [1] "Flow"
#> 
#> $agg_by_colname
#> [1] "Aggregation.by"
#> 
#> $fd_sectors_colname
#> [1] "Fd.sectors"
#> 
#> $p_ind_comp_colname
#> [1] "p_industries_complete"
#> 
#> $p_ind_prefix_colname
#> [1] "p_industry_prefixes"
#> 
#> $ex_colname
#> [1] "EX"
#> 
#> $ex_p_colname
#> [1] "EX.p"
#> 
#> $ex_net_colname
#> [1] "EX.fd_net"
#> 
#> $ex_gross_colname
#> [1] "EX.fd_gross"
#>